Add low-res icon versions and fix install dir
Add low-resolution icons and adjust setup.py script to install
these files to the usual /usr/share/icons/hicolor directory
instead of blaplay/images in dist-packages. Otherwise the GNOME
shell will resort to using an upscaled version of the window icon
in the tab switcher which in turn is a downscaled pixbuf of the
file GTK selects as window icon.