spectrum: check colors in each iteration
Colors in the spectrum visualization were previously only set when they
were known to have changed through the preferences. This, however, prevents
the element from using correct colors if a color change is caused by an
outside change, e.g. when the user changes the GTK theme. We now check in each
call to the draw routine if the colors have changed and update them if necessary.