Set proper font-description for custom cellrenderers
So far subclassed cellrenderers rendering text used the default
font-description, causing all text in the custom library browser and
queue positions in playlists to be of different size possibly. Now the
font-description retrieved from the widget passed to the on_render
method is used.

Additionally, the statusbar text for the queue now displays the correct
cumulative size and length of queued items.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>