blautils.py: catch unhandled exception
The `serialize_to_file' routine uses shutil to back up old files. If the
destination for these backup files exists on the filesytem an exception is
raised which we now deal with properly. Additionally, we now remove backup
files when the actual file is successfully opened in the `deserialize_from_file'
routine.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>