spectrum: protect adapter modifications by lock
Incoming buffers are written and read/flushed from different methods,
executed in different threads. Since the underlying storage type of
buffers in the spectrum visualizer (a GstAdapter) is not thread-safe
modifications to the adapter need to be protected by a lock. This is
likely to fix occasional segfaults that have to do with unreferencing
GstMiniObjects whose reference count already dropped to 0.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>