blatoolbar.py: fix incorrect scroll behaviour
We previously changed BlaPlayer's `get_position' method to return 0 instead
of -1 when querying the bin for the current position fails. This changed was
not applied to the PositionSlider class in blatoolbar.py which still checked
for -1 as error value. On excessive scrolling, causing many seek operations on
the bin, the get_position sometimes returns 0 which the slider class lets pass
as valid value, thus incorrectly resetting the slider position.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>