Simplify singleton check
We only want to allow one instance per user, thus we create a pid file in
the user's blaplay directory on startup. So far, when blaplay was invoked
we checked the directory for a pid file, read the pid and checked the contents
of /proc/$PID which required parsing the cmdline to decide whether to leave
the interpreter or continue. We now create a pidfile in the user directory and
lock it using fcntl, removing the necessity to even read the proc table at all.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>