bladb.py: fix bug with EVENT_MOVED events
Files residing in a monitored directory which are not in the library
(e.g. m3u, png, etc.) would so far raise a KeyError exception in the `move()'
method as it assumed that the method would only be invoked on tracks already in
the library. Consequently, files being renamed to something we might want to
consider parsing based on it's (possibly) new extension were thus ignored. An
example for a situation would be downloading a song through Chrome to a monitored
directory because Chrome appends `crdownload' to downloads that are in progress and
renames the files on completion.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>