Visualization: handle import errors gracefully
When a module raises an exception upon importing we now catch it and just print
the associated message on the console instead of letting the interpreter shut down
due to unhandled exceptions. After all, visualizations aren't vital for the basic
functionality so we might as well keep operating without them.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>