BlaPlayer: don't use playbin2's volume property
Starting with gstreamer 0.10.24 playbin2's volume property and pulseaudio are
linked. That means if pulseaudio is configured to do so its master volume will
follow the loudest element connected to a device mixer. While a good idea at
first glance it frequently causes the system volume to jump from rather low to high
values when volume is adjusted in blaplay. We now circumvent this situation by adding
a separate volume element to our gstreamer pipeline.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>