blautils.py: catch webbrowser exception
Opening a URL in a browser is passed on to the webbrowser module.
While the behaviour is seemingly erratic, if the BROWSER environment
variable is exported and points to a non-existent binary the webbrowser
module occasionally raises an exception (originating from subprocess.py).
We now catch said (OSError) exception just to be on the save side.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>