blabrowsers.py: bugfix in BlaTreeView.__popup_menu
Catch an unhandled exception when trying to open the right-click menu
without having any entries selected in the treeview. Additionally, we
now make the "Open containing directory" option insensitive if not all
selected items reside in the same directory. Lastly, make the option
insensitive in the filesystem browser as well if the selection is not a
directory, and make sure to descend into the actual directory, not the
one containing the selected one.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>