blaplayer.py: handle import errors gracefully
Add dummy class to mimic the basic behaviour of the gst module if it
fails to import. Additionally add a function to verify that gstreamer
works as expected. This is used to inform the user about a missing
gstreamer installation/plugin-set via a message box instead of unhandled
tracebacks on the command line.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>