blaplaylist.py: base shuffle track selection on ID
When selecting a shuffled track from the current playlist via the
`get_random' closure in the `get_track' method we try to make sure that
the last track is not played twice in a row if possible. This could
previously raise an exception if the old track wasn't in the playlist
anymore (due to filtering or removal). However, the way we checked for
duplicates was based on the absolute path in the treemodel which isn't
intuitive to begin with. Instead, we now compare playlist IDs as these
are unique. Additionally, this takes care of the unhandled exception
from before.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>