Prepone parsing of command-line arguments
We recently moved the initialization of the logging module and the
associated modification of __builtin__'s __dict__ up in the
initializiation step for the blaplay module. This caused a regression in
which the logging module would never operate in debug mode as the
command-line option would only be read after the logging module had been
set up. Command-line arguments are now parsed first, then the logging
module is initialized, and only after that do we parse the remaining
command-line arguments.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>