bladb.py: save library on shutdown if necessary
We defer updating the library on disk a few seconds in anticipation of
more changes to avoid writing the library multiple times. The
corresponding URIs in the playlists, however, are updated immediately.
If a shutdown is initiated before the library is updated on disk the new
URIs will be saved in the playlists file, but not in the library file.
On startup this would then raise KeyError exceptions (which we now
catch, by the way). To avoid the problem we now check on shutdown if
there are pending changes to the library that were not written to disk
yet and do so if necessary.