blavisualization.py: fix memory leak
This fixes a bug where the new_buffer method on visualization elements
was still called even after visualizations had been disabled. This was
due to the signal handler of the new_buffer signal emitted by BlaPlayer
not being removed upon disabling of visualizations. As a consequence the
GstAdapter in the visualization element never flushed its buffers,
causing memory use to increase over time.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>