Replace changed URIs properly in playlists
When files in monitored directories change the BlaPlaylist class needs
to update its `uris' dict in order to display the proper metadata from
the library. For that we need to swap keys and values of the dictionary
which previously caused a bug where IDs were dropped due to
non-uniqueness of values in the dict. The new solution uses numpy arrays
for convenience.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>