BlaPlayer: don't use idle_add in message callbacks
The `message' signal callback was previously decorated with blautil.idle
to push the execution to the main thread. According to the docs,
however, the signal on the bus is already emitted from the main thread
making the decorator redundant. Additionally, this commit adds two
context managers around gtk calls which deadlocked blaplay in the past.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>