Don't initialize GTK's threading support
The call to gtk.gdk.threads_init() was removed. This call instructs the gtk
main loop to release the gdk lock after each iteration so it can be
acquired by other threads than the main one. However, since we only
make gtk calls in the main thread anyway this would require us to grab
the lock every time. With this change we don't need to worry about
having the lock at all as the main thread never releases it in the first
place.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>