Remove support for DBus activation again
Support for DBus activation was previously added as it seemed necessary
to allow Cinnamon to start blaplay through its sound menu. However, DBus
activation is not a requirement for this. Cinnamon simply maintains a
hard coded list of supported media players in its sound menu applet.
Being controllable through the sound menu, however, is a matter of
implementing the MPRIS 2 interface.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>