Add support for secondary debugging level
This adds support for controlling the granularity of debugging. Right
now, this is limited to an additional timestamp before the position
of the printed message in the source code. In the future we might use it
for more verbose debugging logs. Additionally, this commit removes a
redundant import of the __builtin__ module as it is already accessible
via the __builtins__ object.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>