blampris.py: fix unintended attribute mutation
D-Bus properties are handled through callbacks stored in a dictionary
which is a class attribute of the MPRIS2 wrapper. The way we implement
the GetAll() D-Bus method would previously override the callback methods
by their return values at the time of execution. Since we returned a
reference to the callback dict from __get_interface_properties() this
would override the callback methods for any subsequent calls as well. We
now define the callback dict as a frozen dict and collect the return
values in the GetAll() method in a separate dictionary.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>