blafm.py: indicate selection in last.fm popup menu
This adds some context info to the last.fm popup menu. When selecting
multiple tracks in a playlist and choosing an option from the last.fm
submenu, it is unclear which playlist entry is affected by the action.
This change adds artist and/or title info to the available options. This
serves as a workaround for #15.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>