Initial version of a major design overhaul
This is the first commit in an attempt to loosen the excessive
coupling between major components of blaplay. Over time, we built up
too many dependencies which made it hard to reason about the effect of
changes to certain components, and which made it largely impossible to
write unit tests. One major issue was the excessive use of
singleton classes throughout the code base. This has been largely
corrected now by using dependency injection in favor of singletons.
There are still a lot of dependency issues to work out, but it's a step
in the right direction. Unfortunately, due to tight coupling, it was too
troublesome to incrementally correct these design issues which is why
this commit is rather invasive, and removes many core features which
will be re-implemented once the new design has been stabilized.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>